# Conflicts:
#	raspd/devtree_quadcopter.lua

+++ raspd/pid.h
@@ -2,27 +2,27 @@
 #define __PID_H__
 
 struct pid_struct {
-    long kp;
-    long ki;
-    long kd;
+    double kp;
+    double ki;
+    double kd;
 
-    long error;
-    long sum_err;
-    long dt_err;
-    long last_input;
-    long min;
-    long max;
-    long output;
+    double error;
+    double sum_err;
+    double dt_err;
+    double last_input;
+    double min;
+    double max;
+    double output;
 };
 
-struct pid_struct *pid_new(long kp, long ki, long kd, long min, long max);
+struct pid_struct *pid_new(double kp, double ki, double kd, double min, double max);
 void pid_del(struct pid_struct *pid);
-long pid_update(struct pid_struct *pid,
-        long setpoint, long input, unsigned long dt);
+double pid_update(struct pid_struct *pid,
+        double setpoint, double input, double dt);
 void pid_set(struct pid_struct *pid,
-        long kp, long ki, long kd, long min, long max);
-void pid_set_tunings(struct pid_struct *pid, long kp, long ki, long kd);
-void pid_set_windup_bounds(struct pid_struct *pid, long min, long max);
+        double kp, double ki, double kd, double min, double max);
+void pid_set_tunings(struct pid_struct *pid, double kp, double ki, double kd);
+void pid_set_windup_bounds(struct pid_struct *pid, double min, double max);
 void pid_reset(struct pid_struct *pid);
 
 #endif /* __PID_H__ */
